[Refactor] enum spells-typeをenum class AttributeTypeに置換
enum classへの置換と必要なキャストを宣言した。
また、実態として呪文の種類を越えて汎用の属性定義になっているため
実態に合わせて名称を更新した。

+++ src/io/interpret-pref-file.cpp
@@ -342,7 +342,7 @@ static errr interpret_z_token(char *buf)
         if (!streq(gf_desc[i].name, buf + 2))
             continue;
 
-        gf_color[gf_desc[i].num] = (TERM_COLOR)quark_add(t);
+        gf_color[(int)gf_desc[i].num] = (TERM_COLOR)quark_add(t);
         return 0;
     }
